Revert "core: make the jailhouse-header arch-independent"
This reverts commit 17f8453876f8b143022278e6438c8ebb157d1821.

It turned out to break 32-bit architectures: At least core_size and
entry must be of the size the hypervisor uses for pointer, otherwise
link-time initialization of those fields fail.

We will need a more sophisticated solution if including the header in
i386 mode is really necessary.

+++ hypervisor/setup.c
@@ -206,5 +206,5 @@ hypervisor_header = {
 	.signature = JAILHOUSE_SIGNATURE,
 	.core_size = (unsigned long)__hv_core_end - JAILHOUSE_BASE,
 	.percpu_size = sizeof(struct per_cpu),
-	.entry = (unsigned long)arch_entry,
+	.entry = arch_entry,
 };
